Update Recommended_Reading.md
Here is a new .md for the assignment, just a little more organized and easy to go through. The one thing i didn't get while doing the readings is why the Data Infographics are included. I'm sure there is a reason that I just don't know, but just saying that going through it I didn't understand the need for it. So maybe that would be something to specify so that the reader would know why they are looking at it, or maybe I just missed it. I will go through the activities later.